requirements.txt
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
imgaug==0.2.9
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.0
opencv-python==4.1.0.25
pdf2image==1.6.0
Pillow==6.1.0
pyparsing==2.4.2
PyPDF2==1.26.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.1
Shapely==1.6.4.post2
six==1.12.0
